+++ controlplane/src/core/bufservices/PlatformService.ts
@@ -2174,6 +2174,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           try {
             const githubRepo = new GitHubRepository(opts.db, opts.githubApp);
             await githubRepo.createCommitCheck({
+              namespace: namespace.name,
               schemaCheckID,
               gitInfo: req.gitInfo,
               compositionErrors,
@@ -2735,19 +2736,9 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           await subgraphRepo.update(
             {
               targetId: subgraph.targetId,
-              labels: req.labels,
-              unsetLabels: req.unsetLabels ?? false,
-              routingUrl: req.routingUrl,
-              subscriptionUrl: req.subscriptionUrl,
+              labels: subgraph.labels,
+              unsetLabels: false,
               schemaSDL: subgraphSchemaSDL,
-              subscriptionProtocol:
-                req.subscriptionProtocol === undefined
-                  ? undefined
-                  : formatSubscriptionProtocol(req.subscriptionProtocol),
-              websocketSubprotocol:
-                req.websocketSubprotocol === undefined
-                  ? undefined
-                  : formatWebsocketSubprotocol(req.websocketSubprotocol),
               updatedBy: authContext.userId,
               namespaceId: namespace.id,
               isV2Graph,
@@ -8019,7 +8010,7 @@ export default function (opts: RouterOptions): Partial<ServiceImpl<typeof Platfo
           search: req.search,
         });
 
-        const count = await orgInvitationRepo.invitationsCount(authContext.organizationId, req.search);
+        const count = await orgInvitationRepo.pendingInvitationsCount(authContext.organizationId, req.search);
 
         return {
           response: {