+++ connect/CHANGELOG.md
@@ -4,6 +4,18 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.72.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-connect@0.72.3...@wundergraph/cosmo-connect@0.72.4) (2024-05-30)
+
+### Bug Fixes
+
+* remove unused check run id causing overflow ([#838](https://github.com/wundergraph/cosmo/issues/838)) ([7fcc5f0](https://github.com/wundergraph/cosmo/commit/7fcc5f06687f5e534ba4056aff4b24d029cd8335)) (@thisisnithin)
+
+## [0.72.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-connect@0.72.2...@wundergraph/cosmo-connect@0.72.3) (2024-05-29)
+
+### Bug Fixes
+
+* prevent subgraph update except schema in publish ([#831](https://github.com/wundergraph/cosmo/issues/831)) ([37a9701](https://github.com/wundergraph/cosmo/commit/37a9701a2b9c61a9ecd489584cd6e2a9fe7ab70b)) (@thisisnithin)
+
 ## [0.72.2](https://github.com/wundergraph/cosmo/compare/@wundergraph/cosmo-connect@0.72.1...@wundergraph/cosmo-connect@0.72.2) (2024-05-24)
 
 ### Bug Fixes