+++ controlplane/CHANGELOG.md
@@ -4,6 +4,40 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.88.10](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.9...controlplane@0.88.10) (2024-06-04)
+
+### Bug Fixes
+
+* missing namespace in github check details link ([#847](https://github.com/wundergraph/cosmo/issues/847)) ([7f413c6](https://github.com/wundergraph/cosmo/commit/7f413c62633cfd082b15e724bb5c9b8118951e99)) (@thisisnithin)
+
+## [0.88.9](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.8...controlplane@0.88.9) (2024-05-31)
+
+### Bug Fixes
+
+* clickhouse cardinality mismatch ([#669](https://github.com/wundergraph/cosmo/issues/669)) ([1908089](https://github.com/wundergraph/cosmo/commit/1908089a6cd62d8e60625555f1173102ce5d8f57)) (@thisisnithin)
+
+## [0.88.8](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.7...controlplane@0.88.8) (2024-05-30)
+
+**Note:** Version bump only for package controlplane
+
+## [0.88.7](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.6...controlplane@0.88.7) (2024-05-29)
+
+### Bug Fixes
+
+* query source table instead of MV to avoid type mismatch ([#837](https://github.com/wundergraph/cosmo/issues/837)) ([a552226](https://github.com/wundergraph/cosmo/commit/a5522269fdebdb8e1e384193075666cb3a8bcc49)) (@StarpTech)
+
+## [0.88.6](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.5...controlplane@0.88.6) (2024-05-29)
+
+### Bug Fixes
+
+* prevent subgraph update except schema in publish ([#831](https://github.com/wundergraph/cosmo/issues/831)) ([37a9701](https://github.com/wundergraph/cosmo/commit/37a9701a2b9c61a9ecd489584cd6e2a9fe7ab70b)) (@thisisnithin)
+
+## [0.88.5](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.4...controlplane@0.88.5) (2024-05-28)
+
+### Bug Fixes
+
+* members invite count and ui spacing ([#833](https://github.com/wundergraph/cosmo/issues/833)) ([487b4e1](https://github.com/wundergraph/cosmo/commit/487b4e18333a0315dca99f1a20884c8bedeff88d)) (@thisisnithin)
+
 ## [0.88.4](https://github.com/wundergraph/cosmo/compare/controlplane@0.88.3...controlplane@0.88.4) (2024-05-24)
 
 ### Bug Fixes